wrong cast return in 64 bit

+++ bundles/org.eclipse.swt/Eclipse SWT PI/cocoa/org/eclipse/swt/internal/cocoa/NSAlert.java
@@ -30,7 +30,7 @@ public static NSAlert alertWithMessageText(NSString message, NSString defaultBut
 }
 
 public int /*long*/ runModal() {
-	return (int)/*64*/OS.objc_msgSend(this.id, OS.sel_runModal);
+	return OS.objc_msgSend(this.id, OS.sel_runModal);
 }
 
 public void setAlertStyle(int /*long*/ style) {
